pyproject.toml
[project]
name = "p0001-samples-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.3"
dependencies = [
    "mcp[cli]>=1.7.1",
    "numpy>=2.2.5",
    "PyJWT>=2.8.0",
    "websockets>=12.0",
    "opentelemetry-sdk>=1.24.0",
    "httpx>=0.28.1",
    "redis>=4.6.0",
    "pandas<=2.2.3",
    "matplotlib>=3.8.4",
    "docker>=7.0.0",
    "PyPDF2>=3.0.1",
    "apscheduler>=3.10.4",
    "sse-starlette>=2.3.4",
]
